+++ choco/tools/chocolateyInstall.ps1
@@ -1,13 +1,14 @@
 $ErrorActionPreference = 'Stop'
-$toolsDir = Split-Path -Parent $MyInvocation.MyCommand.Definition
-$version  = $env:ChocolateyPackageVersion
+$version = $env:ChocolateyPackageVersion
 
 $packageArgs = @{
     packageName    = $env:ChocolateyPackageName
-    fileFullPath   = Join-Path $toolsDir 'KillerShell.exe'
+    fileType       = 'exe'
+    silentArgs     = '/silent'
     url64bit       = "https://github.com/SteveTheKiller/KillerShell/releases/download/v$version/KillerShell.exe"
     checksum64     = 'REPLACE_HASH'
     checksumType64 = 'sha256'
+    validExitCodes = @(0)
 }
 
-Get-ChocolateyWebFile @packageArgs
+Install-ChocolateyPackage @packageArgs